Search Path feature is implemented. Now every search is works fine. User result design should be implemented next.

+++ source/android/Renaissance/lib/models/basic_path.dart
@@ -10,7 +10,7 @@ class BasicPath {
         title: json['title'],
         id: json['_id'],
         photo: json['photo'],
-        effort: json['effort'] as double,
-        rating: json['rating'] as double);
+        effort: json['effort'],
+        rating: json['rating']);
   }
 }
